Fix `(.label)` derived form
Previously `(.label)` derived form generated an expression of the form

    fun (x: _) => x.label

which cannot be typed by the current implementation of 1ML and is unusable as
such.

This commit changes the `(.label)` derived form to generate an expression of the
form

    fun (x: {label: _}) => x.label

which can at least be typed by the current implementation of 1ML and is usable
in some contexts.