Remove `local-in-end` and `include` from the syntax
First of all `include` is superfluous and verbose compared to ellipsis `...`.

Also `local-in-end` is not significantly shorter than its encoding using `...`
and `let-in`:

```diff
- local X in Y end
+ ...let X in {Y}
```

Probably one of the real reasons for having `local-in-end` in Standard ML is the
fact that modules are not first-class in SML.  With first-class modules the
construct is a simple combination of include `...`, `let-in`, and `{-}`.

Given these it seems better to just remove both `local-in-end` and `include` to
simplify the syntax.