Change `if-then-else` for nesting, to omit `else`, not need parens
The implementation introduces a separate `AnnotE` expression.  An `if-then-else`
returning a value of a large type must have its `else` branch explicitly
annotated.  The smart constructor for `IfE` expressions duplicates the
annotation, if any, from the `else` branch to cover the whole `if-then-else`
expression.  This approach has the advantage of keeping the grammar simple.  On
the other hand, this requires minor changes to elaboration.

Additionally the lexer is changed to layout sensitively insert parentheses
around `if-then-else` expressions and insert `else {}` in case an `if-then` ends
without one.

Also made layout sensitive handling of parentheses more precise.