Fix `annotP` such that it always equates the types
Previously `annotP` only equated the types in the case when there was a sequence
of bindings produced by the nested patterns.  This meant that inconsistent
patterns would type check.  For example, before this PR, the REPL interaction

    1ML> ({}: int) = 1;

    1ML> ((x: int): text) = "1";
    x : text

would not give errors.  After this PR, the same interaction

    1ML> ({}: int) = 1;
    stdin:1.2-1.9: inconsistent types equated by `as': () vs int
    1ML> ((x: int): text) = "1";
    stdin:1.6-1.16: inconsistent types equated by `as': int vs text

gives reasonable errors.