Change to layout sensitive syntax
The separator of binds is changed to be `,` instead of `;`.

The syntaxes for `;` separated and `do` expressions are dropped and now the
syntax for expressions, `exp`, includes `let-in-end` expressions.

A layer between the lexer and parser performs layout sensitive insertion of
`,`, `;`, and `in` tokens.

This is fairly similar to the layout sensitive syntax of other FP languages like
F# and Haskell.

The reason for changing the bind separator is that in most cases the separators
should be inserted automatically.  The main exception is cases where a structure
expression is written on a single line, e.g `{x = 1, y = 2}`, and there the use
of `,` instead of `;` looks fine.