Refactor F-omega subtitution and normalization to avoid allocations
Previously the F-omega substitution code was written in a straightforward
manner:

    let rec subst_typ s = function
      (* ... *)
      | PrimT(t) -> PrimT(t)
      | ArrT(t1, t2) -> ArrT(subst_typ s t1, subst_typ s t2)
      (* ... *)

As written, both the `PrimT(t)` and the `ArrT(t1, t2)` always reallocate.

This commit changes the code to explicitly avoid allocations:

    let rec subst_typ s t =
      match t with
      (* ... *)
      | PrimT(_) -> t
      | ArrT(t1, t2) ->
        let t1' = subst_typ s t1 in
        let t2' = subst_typ s t2 in
        if t1 == t1' && t2 == t2' then t else ArrT(t1', t2')
      (* ... *)

The `PrimT(t)` case just reuses the value `t` and the `ArrT(t1, t2)` case reuses
the value in case both of substituted values are (physically) equal to the
original values.

Although the manual optimization does make the code slightly more difficult to
understand, the F-omega checker is a relatively stable part of the 1ML compiler
and these optimizations do give a measurable performance improvement.