Fix typing of structure patterns
Previously structure patterns were typed restrictively such that e.g. in

    {id} = {id}

the `id` function would no longer be pure.  That is because the pattern is
actually elaborated roughly as

    local $ = {id} in id: _ = $.id

and the type ascription `: _` makes the binding of `id` impure.

This changes the typing of patterns such that "inferred" and "annotated" types
for a pattern are maintained separately.  When used as a parameter, both are
used as annotations for the parameter.  When used in a simple binding, the
annotated type is used to constrain the source of the pattern and the inferred
type is only used as a separate constraint on the source that does not restrict
the typing of bindings.